requirements.txt
alembic==0.8.9
click==6.6
et-xmlfile==1.0.1
Flask==0.11.1
Flask-Login==0.4.0
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
gunicorn==19.6.0
itsdangerous==0.24
jdcal==1.3
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
numpy==1.11.2
openpyxl==2.4.1
pandas==0.19.1
psycopg2==2.6.2
pycrypto==2.6.1
python-dateutil==2.6.0
python-dotenv==0.6.1
python-editor==1.0.3
pytz==2016.10
six==1.10.0
SQLAlchemy==1.1.4
Werkzeug==0.11.11
